requirements.txt
Django==5.1.7
djangorestframework==3.14.0
mysqlclient==2.1.0
django-cors-headers>=4.3.1
Pillow==10.0.1
drf-yasg>=1.21.7
